[ENH] refactor probabilistic prediction default dispatching logic to mixin class (#7230)
This PR refactors probabilistic prediction default dispatching logic to
a new mixin class, `_PredictProbaMixin`, to be used with any base class
that implements the probabilistic prediction methods
`_predict_interval`, `_predict_quantiles`, `_predict_var`, and
`_predict_proba`.

This is in preparation for:

* use in the global forecasting interface, see
https://github.com/sktime/sktime/pull/6628
* potential use for time series regression
* potential cross-refactor with `skpro`

The change has as a consequence that the inner `_predict_proba` etc
methods can no longer be called with args (unnamed), though that should
not impact the extension template.
